requirements.txt
cycler==0.10.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.3
openslide-python==1.1.2
Pillow==8.2.0
pkg_resources==0.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
sklearn==0.0
threadpoolctl==2.1.0
torch==1.8.1+cpu
torchaudio==0.8.1
torchvision==0.9.1+cpu
tqdm==4.60.0
typing-extensions==3.10.0.0
